requirements.txt
asgiref==3.8.1
Django==4.2.16
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
joblib==1.4.2
numpy==2.0.2
PyJWT==2.9.0
python-decouple==3.8
scikit-learn==1.5.2
scipy==1.13.1
sqlparse==0.5.1
threadpoolctl==3.5.0
typing_extensions==4.12.2
tzdata==2024.1
